my life

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -6,7 +6,7 @@
  */
 size_t print_list(const list_t *h)
 {
-size_t o = 0;
+int o = 0;
 while (h)
 {
 if (h->str == NULL)
